Added mockito & project structure refactor & mocked FaultyConnection test
