+++ src/test/java/edu/hw2/callerinfo/CallerInfoTest.java
@@ -1,6 +1,5 @@
-package edu.hw2;
+package edu.hw2.callerinfo;
 
-import edu.hw2.callerinfo.CallingInfo;
 import org.junit.jupiter.api.Test;
 import static edu.hw2.callerinfo.CallerInfo.getCallingInfo;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -9,7 +8,7 @@ public class CallerInfoTest {
     @Test
     void callingInfoTest() {
         CallingInfo res = getCallingInfo();
-        CallingInfo expected = new CallingInfo("edu.hw2.CallerInfoTest", "callingInfoTest");
+        CallingInfo expected = new CallingInfo("edu.hw2.callerinfo.CallerInfoTest", "callingInfoTest");
         assertThat(res).isEqualTo(expected);
     }
 }
