+++ src/test/java/edu/hw2/inheritance/InheritanceTest.java
@@ -1,7 +1,5 @@
-package edu.hw2;
+package edu.hw2.inheritance;
 
-import edu.hw2.inheritance.Rectangle;
-import edu.hw2.inheritance.Square;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
